# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/add" : {
      "post" : {
        "summary" : "addCommentLike",
        "description" : "addCommentLike",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "commentId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/sysPermission/add" : {
      "post" : {
        "summary" : "add",
        "description" : "add",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "name", "path", "permission" ],
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "name" : {
                    "type" : "string",
                    "description" : "权限名称"
                  },
                  "permission" : {
                    "type" : "string",
                    "description" : "权限标识 sys:user:list"
                  },
                  "path" : {
                    "type" : "string",
                    "description" : "菜单路由"
                  },
                  "parentId" : {
                    "type" : "integer",
                    "description" : "父菜单ID",
                    "format" : "int64"
                  },
                  "type" : {
                    "type" : "integer",
                    "description" : "1菜单 2按钮"
                  },
                  "children" : {
                    "type" : "array",
                    "description" : "用于封装树形结构的子菜单",
                    "items" : {
                      "required" : [ "name", "path", "permission" ],
                      "type" : "object",
                      "properties" : {
                        "id" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "name" : {
                          "type" : "string",
                          "description" : "权限名称"
                        },
                        "permission" : {
                          "type" : "string",
                          "description" : "权限标识 sys:user:list"
                        },
                        "path" : {
                          "type" : "string",
                          "description" : "菜单路由"
                        },
                        "parentId" : {
                          "type" : "integer",
                          "description" : "父菜单ID",
                          "format" : "int64"
                        },
                        "type" : {
                          "type" : "integer",
                          "description" : "1菜单 2按钮"
                        },
                        "children" : {
                          "type" : "array",
                          "description" : "用于封装树形结构的子菜单",
                          "items" : {
                            "required" : [ "name", "path", "permission" ],
                            "type" : "object",
                            "properties" : {
                              "id" : {
                                "type" : "integer",
                                "description" : "",
                                "format" : "int64"
                              },
                              "name" : {
                                "type" : "string",
                                "description" : "权限名称"
                              },
                              "permission" : {
                                "type" : "string",
                                "description" : "权限标识 sys:user:list"
                              },
                              "path" : {
                                "type" : "string",
                                "description" : "菜单路由"
                              },
                              "parentId" : {
                                "type" : "integer",
                                "description" : "父菜单ID",
                                "format" : "int64"
                              },
                              "type" : {
                                "type" : "integer",
                                "description" : "1菜单 2按钮"
                              },
                              "children" : {
                                "type" : "array",
                                "description" : "用于封装树形结构的子菜单",
                                "items" : {
                                  "type" : "object",
                                  "properties" : { }
                                }
                              }
                            },
                            "description" : "(SysPermission)表实体类"
                          }
                        }
                      },
                      "description" : "(SysPermission)表实体类"
                    }
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : { },
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sysPermission/update" : {
      "put" : {
        "summary" : "update",
        "description" : "update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "name", "path", "permission" ],
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "name" : {
                    "type" : "string",
                    "description" : "权限名称"
                  },
                  "permission" : {
                    "type" : "string",
                    "description" : "权限标识 sys:user:list"
                  },
                  "path" : {
                    "type" : "string",
                    "description" : "菜单路由"
                  },
                  "parentId" : {
                    "type" : "integer",
                    "description" : "父菜单ID",
                    "format" : "int64"
                  },
                  "type" : {
                    "type" : "integer",
                    "description" : "1菜单 2按钮"
                  },
                  "children" : {
                    "type" : "array",
                    "description" : "用于封装树形结构的子菜单",
                    "items" : {
                      "required" : [ "name", "path", "permission" ],
                      "type" : "object",
                      "properties" : {
                        "id" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "name" : {
                          "type" : "string",
                          "description" : "权限名称"
                        },
                        "permission" : {
                          "type" : "string",
                          "description" : "权限标识 sys:user:list"
                        },
                        "path" : {
                          "type" : "string",
                          "description" : "菜单路由"
                        },
                        "parentId" : {
                          "type" : "integer",
                          "description" : "父菜单ID",
                          "format" : "int64"
                        },
                        "type" : {
                          "type" : "integer",
                          "description" : "1菜单 2按钮"
                        },
                        "children" : {
                          "type" : "array",
                          "description" : "用于封装树形结构的子菜单",
                          "items" : {
                            "required" : [ "name", "path", "permission" ],
                            "type" : "object",
                            "properties" : {
                              "id" : {
                                "type" : "integer",
                                "description" : "",
                                "format" : "int64"
                              },
                              "name" : {
                                "type" : "string",
                                "description" : "权限名称"
                              },
                              "permission" : {
                                "type" : "string",
                                "description" : "权限标识 sys:user:list"
                              },
                              "path" : {
                                "type" : "string",
                                "description" : "菜单路由"
                              },
                              "parentId" : {
                                "type" : "integer",
                                "description" : "父菜单ID",
                                "format" : "int64"
                              },
                              "type" : {
                                "type" : "integer",
                                "description" : "1菜单 2按钮"
                              },
                              "children" : {
                                "type" : "array",
                                "description" : "用于封装树形结构的子菜单",
                                "items" : {
                                  "type" : "object",
                                  "properties" : { }
                                }
                              }
                            },
                            "description" : "(SysPermission)表实体类"
                          }
                        }
                      },
                      "description" : "(SysPermission)表实体类"
                    }
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sysPermission/delete" : {
      "delete" : {
        "summary" : "delete",
        "description" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sysPermission/selectOne" : {
      "get" : {
        "summary" : "getById",
        "description" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}